Fix issues with report

+++ generate_report.bash
@@ -4,8 +4,8 @@ OUTPUT_FILE="results/report.json"
 JS_FILE="generate_report.js"
 
 # Axe provides the categories: inapplicable, passes, incomplete, violation
-# Axes fail condition: violation (non customizable)
-# Axes reports: inapplicable, passes, incomplete, violation (non customizable)
+# Axe fail condition: violation (non customizable)
+# Axe reports: inapplicable, passes, incomplete, violation (non customizable)
 
 # Achecker provides the categories: "violation", "potentialviolation","recommendation","potentialrecommendation",
 # "manual","pass","ignored","elements","elementsViolation","elementsViolationReview"
@@ -22,24 +22,18 @@ function axe-results-summary {
 
     # Check if the target is not a directory
     if [ ! -d "$axe_directory" ]; then
-    echo "Not a directory."
+    echo "The '$axe_directory' directory does not exist."
     exit 1
     fi
 
     # Loop through files in the target directory
     for file in "$axe_directory"/*; do
     if [ -f "$file" ]; then
-        # url examined
-        url=$(jq '.[0].url' "$file")
-        # total errors for url
-        # total_errors=$(jq '[.[0].violations[].nodes | length | tonumber ] | add | if . == null then 0 else . end' "$file")
-        total_errors=$(jq '.[].violations | length' "$file")
-        echo "Axe run on $url found $total_errors error(s)."
-        node $JS_FILE "./$file"
+        node "$JS_FILE" "axe" "./$file"
     fi
     done    
     
-    echo "---------------------------------------------"
+    echo "Axe results report is done."
 }
 
 function achecker-results-summary {
@@ -48,25 +42,25 @@ function achecker-results-summary {
 
     # Check if the target is not a directory
     if [ ! -d "$achecker_directory" ]; then
-    echo "Not a directory."
+    echo "The '$achecker_directory' directory does not exist."
     exit 1
     fi
 
     # Rename home page file
     # Needed for being read in the loop below! 
-    mv "$achecker_directory/.json" "$achecker_directory/home.json"
+    if [ -e "$achecker_directory/.json" ]; then
+        mv "$achecker_directory/.json" "$achecker_directory/home.json"
+    fi
+    
 
     # Loop through files in the target directory
     for file in "$achecker_directory"/*; do
     if [ -f "$file" ]; then
-        url=$(jq '.summary.URL' "$file")
-        total_errors=$(jq '.summary.counts.violation' "$file")
-        echo "Achecker run on $url found $total_errors error(s)."
-        node $JS_FILE "./$file"
+        node "$JS_FILE" "achecker" "./$file"
     fi
     done
 
-    echo "---------------------------------------------"
+    echo "Accessibility-checker results report is done."
 }
 
 function htmlcs-results-summary {
@@ -75,26 +69,27 @@ function htmlcs-results-summary {
 
     # Check if the target is not a directory
     if [ ! -d "$htmlcs_directory" ]; then
-    echo "Not a directory."
+    echo "The '$htmlcs_directory' directory does not exist."
     exit 1
     fi
 
     # Loop through files in the target directory
     for file in "$htmlcs_directory"/*; do
     if [ -f "$file" ]; then
-        # url examined
-        url=$( head -n 1 "$file" )
-        # # total numbers of errors
-        total_errors=$( grep -wc "Error" "$file")
-        echo "HTML_CodeSniffer run on $url found $total_errors error(s)."
-        node $JS_FILE "./$file"
+        node "$JS_FILE" "htmlcs" "./$file"
     fi
     done
 
-    echo "---------------------------------------------"
+    echo "HTML_CodeSniffer results report is done."
+
 }
 
-echo '{"summary": [], "report": []}' > $OUTPUT_FILE
-achecker-results-summary
-axe-results-summary
-htmlcs-results-summary
+function main {
+    # Create output file to write to
+    echo '{"summary": [], "report": []}' > $OUTPUT_FILE
+    achecker-results-summary
+    axe-results-summary
+    htmlcs-results-summary
+}
+
+main
